+++ cmd/update/app/flag.go
@@ -9,16 +9,19 @@ import (
 const (
 	flagVersion = "version"
 	flagName    = "name"
+	flagSuspend = "suspend-reconciliation"
 )
 
 type flag struct {
-	print   *genericclioptions.PrintFlags
-	Name    string
-	Version string
+	print                 *genericclioptions.PrintFlags
+	Name                  string
+	SuspendReconciliation bool
+	Version               string
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.Version, flagVersion, "", "Version to update the app to")
+	cmd.Flags().BoolVar(&f.SuspendReconciliation, flagSuspend, false, "Suspend app reconciliation by Flux")
 	// Hide flag in favour of the longDescription, otherwise if the number of supported
 	// update flags grows, it may be hard to differentiate them from the rest of the flags,
 	// like kubectl global flags.
@@ -37,10 +40,6 @@ func (f *flag) Init(cmd *cobra.Command) {
 func (f *flag) Validate() error {
 	if f.Name == "" {
 		return microerror.Maskf(invalidFlagError, "--%s must not be empty", flagName)
-	}
-
-	if f.Version == "" {
-		return microerror.Maskf(invalidFlagError, "--%s must not be empty", flagVersion)
 	}
 
 	return nil