+++ cmd/update/app/runner_test.go
@@ -49,7 +49,7 @@ func Test_run(t *testing.T) {
 				newAppCatalogEntry("fake-app", "0.1.0", "fake-catalog", "true"),
 			},
 			flags:   flag{Name: "fake-app", Version: "0.1.0"},
-			message: "App 'fake-app' updated to version '0.1.0'\n",
+			message: "App \"fake-app\" in namespace \"default\" updated with version=0.1.0\n",
 		},
 		{
 			name: "patch app with the AppCatalogEntry CR (not latest)",
@@ -61,7 +61,7 @@ func Test_run(t *testing.T) {
 				newAppCatalogEntry("fake-app", "0.2.0", "fake-catalog", "true"),
 			},
 			flags:   flag{Name: "fake-app", Version: "0.1.0"},
-			message: "App 'fake-app' updated to version '0.1.0'\n",
+			message: "App \"fake-app\" in namespace \"default\" updated with version=0.1.0\n",
 		},
 		{
 			name: "patch app without AppCatalogEntry CR, but available in catalog",
@@ -76,7 +76,7 @@ func Test_run(t *testing.T) {
 			},
 			flags:             flag{Name: "fake-app", Version: "0.0.1"},
 			chartResponseCode: 200,
-			message:           "App 'fake-app' updated to version '0.0.1'\n",
+			message:           "App \"fake-app\" in namespace \"default\" updated with version=0.0.1\n",
 		},
 		{
 			name: "patch app with nonexisting version",